+++ Futebol.Domain/Commands/ConsultarTimeCommand.cs
@@ -1,11 +1,12 @@
+using Futebol.Domain.Commands.Responses;
 using MediatR;
 
 namespace Futebol.Domain.Commands
 {
     public class ConsultarTimeCommand : IRequest<ConsultarTimeResponse>
     {
-        public long Id { get; init; }
-        public ConsultarTimeCommand(long id) {
+        public int Id { get; init; }
+        public ConsultarTimeCommand(int id) {
             Id = id;
         }
     }
