+++ Futebol.Domain/Repositories/ITimeRepository.cs
@@ -5,7 +5,7 @@ namespace Futebol.Domain.Repositories
     public interface ITimeRepository
     {
         Task<IEnumerable<Time>> ObterTodosAsync();
-        Task<Time> ObterPorIdAsync(long id);
+        Task<Time> ObterPorIdAsync(int id);
         Task<Time> ObterPorNomeAsync(string nome);
         Task CriarAsync(Time time);
         Task AtualizarAsync(Time time);
